rule UnionRewrite
