+++ scalafix/.scalafix.conf
@@ -1,3 +1,3 @@
-SparkAutoUpgrade.deprecatedMethod {
+UnionRewrite.deprecatedMethod {
   "unionAll" = "union"
 }
