+++ scalafix/rules/src/main/resources/META-INF/services/scalafix.v1.Rule
@@ -5,3 +5,4 @@ fix.MigrateDeprecatedDataFrameReaderFuns
 fix.AccumulatorUpgrade
 fix.OnFailureFix
 fix.ExecutorPluginWarn
+fix.UnionRewrite
